Split out ANYTHING_ELSE, move Charclass Fsm constructor
